Colección de notas.

Contiene colecciones de notas de distintos portales noticiosos (una muestra pequeña). Una segunda colección es de notas del periódico bahiense La Nueva. También tiene resúmenes estadísticos de las bases completas para el desarrollo de los ejemplos de las funciones.

data(acep_bases)

Format

Es una lista con 8 objetos.

la_nueva

es un data frame con notas de La Nueva

rev_puerto

es un data frame con notas de la Revista Puerto

rp_procesada

es un data frame con indicadores de conflictividad basados en los datos de la Revista Puerto

lc_mdp

es una url para la descarga del corpus de notas de La Capital

rp_mdp

es una url para la descarga del corpus de notas de la Revista Puerto

ed_neco

es una url para la descarga del corpus de notas de Ecos Diarios

ln_bb

es una url para la descarga del corpus de notas de La Nueva

ln_arg

es una url para la descarga del corpus de notas de La Nación

Source

Revista Puerto

La Nueva

References

Nieto, Agustín 2020 «Intersecciones entre historia digital e historia social: un ejercicio de lectura distante sobre la conflictividad marítima en la historia argentina reciente». Drassana: revista del Museu Maritim (28):122-42. (Revista Drassana)

Examples

la_nueva <- acep_bases$la_nueva
head(la_nueva, 10)
#> # A tibble: 10 × 3
#>    fecha      titulo                                                       nota 
#>    <date>     <chr>                                                        <chr>
#>  1 2022-01-15 "Llevan 36 horas sin energía por un cable suelto y EDES cal… "Vec…
#>  2 2022-01-15 "Explotó un generador que EDES colocó en el centro de la ci… "Mom…
#>  3 2022-01-15 "Amenazaron a trabajadores de ABSA en el barrio Stella Mari… "Tra…
#>  4 2022-01-15 "Intenso calor y falta de lluvia, el combo ideal para los i… "Las…
#>  5 2022-01-15 "Investigan un incendio que destruyó un automóvil"           "La …
#>  6 2022-01-15 "Con la ola de contagios, ¿cuándo vuelven las clases?"       "Por…
#>  7 2022-01-15 "Pavimento que estalla: ¿a qué se deben las roturas?"        "Aud…
#>  8 2022-01-15 "¿Cuánto falta para que la pandemia llegue a su fin y pase … "Aud…
#>  9 2022-01-14 "Vecinos del barrio Patagonia piden que avance una obra en … "La …
#> 10 2022-01-14 "Otra renuncia en el Municipio: se va Karina Mahon"          "La …