Frecuencia, menciones e intensidad.

Función que usa las funciones acep_frec, acep_men y acep_int y devuelve una tabla con tres columnas nuevas: número de palabras, número de menciones del diccionario, índice de intensidad.

acep_db(db, t, d, n)

Arguments

db

data frame con los textos a procesar.

t

columna de data frame que contiene el vector de textos a procesar.

d

diccionario en formato vector.

n

cantidad de decimales del índice de intensidad.

Value

Si todas las entradas son correctas, la salida será una base de datos en formato tabular con tres nuevas variables.

Examples

rev_puerto <- acep_bases$rev_puerto
dicc_viol_gp <- acep_diccionarios$dicc_viol_gp
rpp <- acep_db(rev_puerto, rev_puerto$nota, dicc_viol_gp, 4)
rpp |> head()
#> # A tibble: 6 × 6
#>   fecha      titulo                       nota  n_palabras conflictos intensidad
#>   <date>     <chr>                        <chr>      <int>      <int>      <dbl>
#> 1 2019-03-29 Astillero Contessi botó al … "Tra…        681          1     0.0015
#> 2 2019-03-29 Di Leva despidió a trabajad… "Tra…        318          0     0     
#> 3 2019-03-29 El Consejo aprobó la apertu… "En …        705          1     0.0014
#> 4 2019-03-28 “Todos los empresarios me d… "Cri…       1118          0     0     
#> 5 2019-03-28 En el Puerto de Montevideo … "La …        952          3     0.0032
#> 6 2019-03-28 Habilitan el muelle Piedra … "La …        705          0     0