Colección de diccionarios.

Colección de diccionarios que reúne diccionarios de diferentes orígenes. El diccionario dicc_confl_sismos fue construido en el marco del Observatorio de Conflictividad de la UNMdP. Los diccionarios dicc_confl_gp y dicc_viol_gp fueron extraídos de Albrieu y Palazzo (2020).

Usage

data(acep_diccionarios)

Format

Es un objeto de clase 'list' con 3 componentes. Cada componente es una cadena de caracteres de longitud 1 con la URL de un diccionario `.rds` descargable; usá acep_load_base para obtener el vector de palabras correspondiente.

dicc_confl_gp

URL al diccionario de términos que refieren a conflictos

dicc_viol_gp

URL al diccionario de términos que refieren a violencia

dicc_confl_sismos

URL al diccionario de términos que refieren a conflictos

Source

Revista Puerto

La Nueva

References

Albrieu, Ramiro y Gabriel Palazzo 2020 «Categorización de conflictos sociales en el ámbito de los recursos naturales: un estudio de las actividades extractivas mediante la minería de textos». Revista CEPAL (131):29-59. (Revista CEPAL)

Laitano, Guillermina y Agustín Nieto «Análisis computacional de la conflictividad laboral en Mar del Plata durante el gobierno de Cambiemos». Ponencia presentado en VI Workshop - Los conflictos laborales en la Argentina del siglo XX y XXI: un abordaje interdisciplinario de conceptos, problemas y escalas de análisis, Tandil, 2021.

Examples

if (FALSE) { # \dontrun{
diccionario <- acep_load_base(acep_diccionarios$dicc_viol_gp)
diccionario
} # }