Frecuencia de palabras totales.

Función que cuenta la frecuencia de palabras totales en cada una de las notas/textos.

acep_frec(x)

Arguments

x

vector de textos al que se le aplica la función de conteo de la frecuencia de palabras.

Value

Si todas las entradas son correctas, la salida será un vector con una frecuencia de palabras.

Examples

rev_puerto <- rev_puerto <- acep_bases$rev_puerto
rev_puerto$n_palabras <- acep_frec(rev_puerto$nota)
rev_puerto |> head()
#> # A tibble: 6 × 4
#>   fecha      titulo                                             nota  n_palabras
#>   <date>     <chr>                                              <chr>      <int>
#> 1 2019-03-29 Astillero Contessi botó al Luca Santino, un fresq… "Tra…        681
#> 2 2019-03-29 Di Leva despidió a trabajadores de El Marisco y S… "Tra…        318
#> 3 2019-03-29 El Consejo aprobó la apertura al norte para el ca… "En …        705
#> 4 2019-03-28 “Todos los empresarios me dicen que tendrían que … "Cri…       1118
#> 5 2019-03-28 En el Puerto de Montevideo siguen bajando un muer… "La …        952
#> 6 2019-03-28 Habilitan el muelle Piedra Buena para descarga de… "La …        705