Índice de intensidad.

Función que elabora un indice de intensidad en base a la relación entre palabras totales y palabras del diccionario presentes en el texto.

Usage

acep_int(pc, pt, decimales = 4)

Arguments

pc

vector numérico con la frecuencia de palabras conflictivas presentes en cada texto.

pt

vector de palabras totales en cada texto.

decimales

cantidad de decimales, por defecto tiene 4 pero se puede modificar.

Value

Si todas las entradas son correctas, la salida será un vector numérico.

Examples

rev_puerto <- acep_bases$rev_puerto
dicc_violencia <- acep_diccionarios$dicc_viol_gp
rev_puerto$n_palabras <- acep_frec(rev_puerto$nota)
rev_puerto$conflictos <- acep_men(rev_puerto$nota, dicc_violencia)
rev_puerto$intensidad <- acep_int(rev_puerto$conflictos, rev_puerto$n_palabras, 3)
rev_puerto |> head()
#> # A tibble: 6 × 6
#>   fecha      titulo                       nota  n_palabras conflictos intensidad
#>   <date>     <chr>                        <chr>      <int>      <int>      <dbl>
#> 1 2019-03-29 Astillero Contessi botó al … "Tra…        681          1      0.001
#> 2 2019-03-29 Di Leva despidió a trabajad… "Tra…        318          0      0    
#> 3 2019-03-29 El Consejo aprobó la apertu… "En …        705          1      0.001
#> 4 2019-03-28 “Todos los empresarios me d… "Cri…       1118          0      0    
#> 5 2019-03-28 En el Puerto de Montevideo … "La …        952          3      0.003
#> 6 2019-03-28 Habilitan el muelle Piedra … "La …        705          0      0