Tokenizador.

Función que tokeniza las notas/textos.

Usage

acep_token(x, tolower = TRUE)

Arguments

x

vector de textos al que se le aplica la función de tokenización.

tolower

convierte los textos a minúsculas.

Examples

rev_puerto <- acep_bases$rev_puerto
rev_puerto_token <- acep_token(rev_puerto$nota)
rev_puerto_token |> head()
#>   id_doc
#> 1      1
#> 2      1
#> 3      1
#> 4      1
#> 5      1
#> 6      1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   texto
#> 1 Tras un año de intento trabajo y una inversión de 3 millones de dólares, este jueves Astillero Contessi entregó su obra número 130. El buque fresquero de altura Luca Santino, ordenado por la pesquera Cabo Vírgenes, bien podría representar un mensaje para todos los actores involucrados en la industria pesquera.\nEl buque de 26,90 metros de eslora y doble cubierta cuenta con una máquina de hielo líquido que por primera vez se instala en un barco pesquero de la flota nacional. La innovación de Eduardo del Río fue destacada por Domingo Contessi desde el discurso previo a la botadura. Resignar bodega para instalar el dispositivo es un paso adelante hacia priorizar la calidad por sobre la cantidad.\nLa empresa armadora, encabezada por Del Río, reprocesa langostino en una planta de Rawson y hace poco adquirió dos buques fresqueros usados, casualmente, construidos en Contessi. Con financiamiento propio, ahora incorporan un nuevo buque para sumar volumen a la línea de producción, pero fundamentalmente, calidad, a partir de la posibilidad de obtener una mayor proporción de langostino entero en condiciones óptimas de frescura.\nEl particular contexto que atraviesa la industria naval post Decreto 145 ocupó buena parte del discurso de Domingo Contessi. «Estamos a las puertas de un potencial resurgimiento de la Industria Naval de la mano de la pesca. Todos los astilleros del país nos sentimos comprometidos con este objetivo y dispuestos a invertir para estar a la altura de las circunstancias”, dijo el industrial en representación del Directorio.\nAntes de ingresar al astillero se observan los movimientos de obra en el nuevo predio gestionado ante el Consorcio Portuario donde planifican la ampliación para poder construir barcos más grandes que el Luca Santino.\nYa en el interior de la nave central se descubren en distintos grados de avance las nuevas construcciones planificadas para ser botadas este mismo año. La última, la 135, es otro barco para Cabo Vírgenes, el Espartano, un fresquero de 21 metros, también con innovaciones en su línea y equipamiento.\nTodo este movimiento generó que el astillero amplíe su planilla de personal un 60% y haya incorporado nuevas maquinarias para hacer frente a la mayor demanda de trabajo.\nLa ampliación se realiza sobre un terreno fiscal y con fondos propios. “Es difícil no pensar en estos momentos que hace apenas 6 años fuimos intimados a la restitución del predio por un interventor del Puerto de Mar del Plata”, dijo el directivo, en tanto reconoció que “ni aún en esos aciagos momentos se nos ocurrió protestar cortando los accesos al puerto o afectando el trabajo de otros, buscamos formas diferentes de visibilizar nuestro reclamo, litigamos judicialmente y gestionamos”, dijo Contessi.\nPareció un mensaje a los directivos de TC2 y la cooperativa Hipocoop, cuyos trabajadores y otros estibadores y transportistas, bloquearon todos los accesos la semana pasada y fue necesaria la intervención de Tizado.\n“Hoy debemos agradecer al Presidente y Directorio del Consorcio Portuario, porque después de 22 años de precariedad tenemos nuestro permiso de uso de suelo renovado y accedimos a un aumento de la superficie ocupada que nos permite concretar esta primera etapa de la ampliación edilicia”.\nContessi también planteó los peligros que merodean al desarrollo pleno de la industria naval. ”Hay sectores que pujan por una rebaja arancelaria para la importación de determinados segmentos de buques nuevos y usados. En la letra chica de esa medida, en una segmentación que no excluya a barcos como el de 40 metros que está en construcción dentro de nuestros galpones y en las excepciones absolutamente innecesarias a la normativa vigente, podría irse por la borda todo lo bueno que se ha hecho en estos últimos meses”, reconoció el Director.\nEntre las autoridades presentes en la botadura estuvieron el ministro de la Producción, Javier Tizado, quien destacó el trabajo del astillero y su apuesta por la generación de más empleo, inversión y desarrollo en tanto recordó el último bloqueo que sufrió el puerto y adujo problemas “de comunicación”. “El puerto es de ustedes”, les dijo a los presentes, entre los que había muchos actores pesqueros. “El sector de la pesca crece fuertemente y vamos a acompañar su desarrollo”, dijo el Ministro.\n    
#> 2 Tras un año de intento trabajo y una inversión de 3 millones de dólares, este jueves Astillero Contessi entregó su obra número 130. El buque fresquero de altura Luca Santino, ordenado por la pesquera Cabo Vírgenes, bien podría representar un mensaje para todos los actores involucrados en la industria pesquera.\nEl buque de 26,90 metros de eslora y doble cubierta cuenta con una máquina de hielo líquido que por primera vez se instala en un barco pesquero de la flota nacional. La innovación de Eduardo del Río fue destacada por Domingo Contessi desde el discurso previo a la botadura. Resignar bodega para instalar el dispositivo es un paso adelante hacia priorizar la calidad por sobre la cantidad.\nLa empresa armadora, encabezada por Del Río, reprocesa langostino en una planta de Rawson y hace poco adquirió dos buques fresqueros usados, casualmente, construidos en Contessi. Con financiamiento propio, ahora incorporan un nuevo buque para sumar volumen a la línea de producción, pero fundamentalmente, calidad, a partir de la posibilidad de obtener una mayor proporción de langostino entero en condiciones óptimas de frescura.\nEl particular contexto que atraviesa la industria naval post Decreto 145 ocupó buena parte del discurso de Domingo Contessi. «Estamos a las puertas de un potencial resurgimiento de la Industria Naval de la mano de la pesca. Todos los astilleros del país nos sentimos comprometidos con este objetivo y dispuestos a invertir para estar a la altura de las circunstancias”, dijo el industrial en representación del Directorio.\nAntes de ingresar al astillero se observan los movimientos de obra en el nuevo predio gestionado ante el Consorcio Portuario donde planifican la ampliación para poder construir barcos más grandes que el Luca Santino.\nYa en el interior de la nave central se descubren en distintos grados de avance las nuevas construcciones planificadas para ser botadas este mismo año. La última, la 135, es otro barco para Cabo Vírgenes, el Espartano, un fresquero de 21 metros, también con innovaciones en su línea y equipamiento.\nTodo este movimiento generó que el astillero amplíe su planilla de personal un 60% y haya incorporado nuevas maquinarias para hacer frente a la mayor demanda de trabajo.\nLa ampliación se realiza sobre un terreno fiscal y con fondos propios. “Es difícil no pensar en estos momentos que hace apenas 6 años fuimos intimados a la restitución del predio por un interventor del Puerto de Mar del Plata”, dijo el directivo, en tanto reconoció que “ni aún en esos aciagos momentos se nos ocurrió protestar cortando los accesos al puerto o afectando el trabajo de otros, buscamos formas diferentes de visibilizar nuestro reclamo, litigamos judicialmente y gestionamos”, dijo Contessi.\nPareció un mensaje a los directivos de TC2 y la cooperativa Hipocoop, cuyos trabajadores y otros estibadores y transportistas, bloquearon todos los accesos la semana pasada y fue necesaria la intervención de Tizado.\n“Hoy debemos agradecer al Presidente y Directorio del Consorcio Portuario, porque después de 22 años de precariedad tenemos nuestro permiso de uso de suelo renovado y accedimos a un aumento de la superficie ocupada que nos permite concretar esta primera etapa de la ampliación edilicia”.\nContessi también planteó los peligros que merodean al desarrollo pleno de la industria naval. ”Hay sectores que pujan por una rebaja arancelaria para la importación de determinados segmentos de buques nuevos y usados. En la letra chica de esa medida, en una segmentación que no excluya a barcos como el de 40 metros que está en construcción dentro de nuestros galpones y en las excepciones absolutamente innecesarias a la normativa vigente, podría irse por la borda todo lo bueno que se ha hecho en estos últimos meses”, reconoció el Director.\nEntre las autoridades presentes en la botadura estuvieron el ministro de la Producción, Javier Tizado, quien destacó el trabajo del astillero y su apuesta por la generación de más empleo, inversión y desarrollo en tanto recordó el último bloqueo que sufrió el puerto y adujo problemas “de comunicación”. “El puerto es de ustedes”, les dijo a los presentes, entre los que había muchos actores pesqueros. “El sector de la pesca crece fuertemente y vamos a acompañar su desarrollo”, dijo el Ministro.\n    
#> 3 Tras un año de intento trabajo y una inversión de 3 millones de dólares, este jueves Astillero Contessi entregó su obra número 130. El buque fresquero de altura Luca Santino, ordenado por la pesquera Cabo Vírgenes, bien podría representar un mensaje para todos los actores involucrados en la industria pesquera.\nEl buque de 26,90 metros de eslora y doble cubierta cuenta con una máquina de hielo líquido que por primera vez se instala en un barco pesquero de la flota nacional. La innovación de Eduardo del Río fue destacada por Domingo Contessi desde el discurso previo a la botadura. Resignar bodega para instalar el dispositivo es un paso adelante hacia priorizar la calidad por sobre la cantidad.\nLa empresa armadora, encabezada por Del Río, reprocesa langostino en una planta de Rawson y hace poco adquirió dos buques fresqueros usados, casualmente, construidos en Contessi. Con financiamiento propio, ahora incorporan un nuevo buque para sumar volumen a la línea de producción, pero fundamentalmente, calidad, a partir de la posibilidad de obtener una mayor proporción de langostino entero en condiciones óptimas de frescura.\nEl particular contexto que atraviesa la industria naval post Decreto 145 ocupó buena parte del discurso de Domingo Contessi. «Estamos a las puertas de un potencial resurgimiento de la Industria Naval de la mano de la pesca. Todos los astilleros del país nos sentimos comprometidos con este objetivo y dispuestos a invertir para estar a la altura de las circunstancias”, dijo el industrial en representación del Directorio.\nAntes de ingresar al astillero se observan los movimientos de obra en el nuevo predio gestionado ante el Consorcio Portuario donde planifican la ampliación para poder construir barcos más grandes que el Luca Santino.\nYa en el interior de la nave central se descubren en distintos grados de avance las nuevas construcciones planificadas para ser botadas este mismo año. La última, la 135, es otro barco para Cabo Vírgenes, el Espartano, un fresquero de 21 metros, también con innovaciones en su línea y equipamiento.\nTodo este movimiento generó que el astillero amplíe su planilla de personal un 60% y haya incorporado nuevas maquinarias para hacer frente a la mayor demanda de trabajo.\nLa ampliación se realiza sobre un terreno fiscal y con fondos propios. “Es difícil no pensar en estos momentos que hace apenas 6 años fuimos intimados a la restitución del predio por un interventor del Puerto de Mar del Plata”, dijo el directivo, en tanto reconoció que “ni aún en esos aciagos momentos se nos ocurrió protestar cortando los accesos al puerto o afectando el trabajo de otros, buscamos formas diferentes de visibilizar nuestro reclamo, litigamos judicialmente y gestionamos”, dijo Contessi.\nPareció un mensaje a los directivos de TC2 y la cooperativa Hipocoop, cuyos trabajadores y otros estibadores y transportistas, bloquearon todos los accesos la semana pasada y fue necesaria la intervención de Tizado.\n“Hoy debemos agradecer al Presidente y Directorio del Consorcio Portuario, porque después de 22 años de precariedad tenemos nuestro permiso de uso de suelo renovado y accedimos a un aumento de la superficie ocupada que nos permite concretar esta primera etapa de la ampliación edilicia”.\nContessi también planteó los peligros que merodean al desarrollo pleno de la industria naval. ”Hay sectores que pujan por una rebaja arancelaria para la importación de determinados segmentos de buques nuevos y usados. En la letra chica de esa medida, en una segmentación que no excluya a barcos como el de 40 metros que está en construcción dentro de nuestros galpones y en las excepciones absolutamente innecesarias a la normativa vigente, podría irse por la borda todo lo bueno que se ha hecho en estos últimos meses”, reconoció el Director.\nEntre las autoridades presentes en la botadura estuvieron el ministro de la Producción, Javier Tizado, quien destacó el trabajo del astillero y su apuesta por la generación de más empleo, inversión y desarrollo en tanto recordó el último bloqueo que sufrió el puerto y adujo problemas “de comunicación”. “El puerto es de ustedes”, les dijo a los presentes, entre los que había muchos actores pesqueros. “El sector de la pesca crece fuertemente y vamos a acompañar su desarrollo”, dijo el Ministro.\n    
#> 4 Tras un año de intento trabajo y una inversión de 3 millones de dólares, este jueves Astillero Contessi entregó su obra número 130. El buque fresquero de altura Luca Santino, ordenado por la pesquera Cabo Vírgenes, bien podría representar un mensaje para todos los actores involucrados en la industria pesquera.\nEl buque de 26,90 metros de eslora y doble cubierta cuenta con una máquina de hielo líquido que por primera vez se instala en un barco pesquero de la flota nacional. La innovación de Eduardo del Río fue destacada por Domingo Contessi desde el discurso previo a la botadura. Resignar bodega para instalar el dispositivo es un paso adelante hacia priorizar la calidad por sobre la cantidad.\nLa empresa armadora, encabezada por Del Río, reprocesa langostino en una planta de Rawson y hace poco adquirió dos buques fresqueros usados, casualmente, construidos en Contessi. Con financiamiento propio, ahora incorporan un nuevo buque para sumar volumen a la línea de producción, pero fundamentalmente, calidad, a partir de la posibilidad de obtener una mayor proporción de langostino entero en condiciones óptimas de frescura.\nEl particular contexto que atraviesa la industria naval post Decreto 145 ocupó buena parte del discurso de Domingo Contessi. «Estamos a las puertas de un potencial resurgimiento de la Industria Naval de la mano de la pesca. Todos los astilleros del país nos sentimos comprometidos con este objetivo y dispuestos a invertir para estar a la altura de las circunstancias”, dijo el industrial en representación del Directorio.\nAntes de ingresar al astillero se observan los movimientos de obra en el nuevo predio gestionado ante el Consorcio Portuario donde planifican la ampliación para poder construir barcos más grandes que el Luca Santino.\nYa en el interior de la nave central se descubren en distintos grados de avance las nuevas construcciones planificadas para ser botadas este mismo año. La última, la 135, es otro barco para Cabo Vírgenes, el Espartano, un fresquero de 21 metros, también con innovaciones en su línea y equipamiento.\nTodo este movimiento generó que el astillero amplíe su planilla de personal un 60% y haya incorporado nuevas maquinarias para hacer frente a la mayor demanda de trabajo.\nLa ampliación se realiza sobre un terreno fiscal y con fondos propios. “Es difícil no pensar en estos momentos que hace apenas 6 años fuimos intimados a la restitución del predio por un interventor del Puerto de Mar del Plata”, dijo el directivo, en tanto reconoció que “ni aún en esos aciagos momentos se nos ocurrió protestar cortando los accesos al puerto o afectando el trabajo de otros, buscamos formas diferentes de visibilizar nuestro reclamo, litigamos judicialmente y gestionamos”, dijo Contessi.\nPareció un mensaje a los directivos de TC2 y la cooperativa Hipocoop, cuyos trabajadores y otros estibadores y transportistas, bloquearon todos los accesos la semana pasada y fue necesaria la intervención de Tizado.\n“Hoy debemos agradecer al Presidente y Directorio del Consorcio Portuario, porque después de 22 años de precariedad tenemos nuestro permiso de uso de suelo renovado y accedimos a un aumento de la superficie ocupada que nos permite concretar esta primera etapa de la ampliación edilicia”.\nContessi también planteó los peligros que merodean al desarrollo pleno de la industria naval. ”Hay sectores que pujan por una rebaja arancelaria para la importación de determinados segmentos de buques nuevos y usados. En la letra chica de esa medida, en una segmentación que no excluya a barcos como el de 40 metros que está en construcción dentro de nuestros galpones y en las excepciones absolutamente innecesarias a la normativa vigente, podría irse por la borda todo lo bueno que se ha hecho en estos últimos meses”, reconoció el Director.\nEntre las autoridades presentes en la botadura estuvieron el ministro de la Producción, Javier Tizado, quien destacó el trabajo del astillero y su apuesta por la generación de más empleo, inversión y desarrollo en tanto recordó el último bloqueo que sufrió el puerto y adujo problemas “de comunicación”. “El puerto es de ustedes”, les dijo a los presentes, entre los que había muchos actores pesqueros. “El sector de la pesca crece fuertemente y vamos a acompañar su desarrollo”, dijo el Ministro.\n    
#> 5 Tras un año de intento trabajo y una inversión de 3 millones de dólares, este jueves Astillero Contessi entregó su obra número 130. El buque fresquero de altura Luca Santino, ordenado por la pesquera Cabo Vírgenes, bien podría representar un mensaje para todos los actores involucrados en la industria pesquera.\nEl buque de 26,90 metros de eslora y doble cubierta cuenta con una máquina de hielo líquido que por primera vez se instala en un barco pesquero de la flota nacional. La innovación de Eduardo del Río fue destacada por Domingo Contessi desde el discurso previo a la botadura. Resignar bodega para instalar el dispositivo es un paso adelante hacia priorizar la calidad por sobre la cantidad.\nLa empresa armadora, encabezada por Del Río, reprocesa langostino en una planta de Rawson y hace poco adquirió dos buques fresqueros usados, casualmente, construidos en Contessi. Con financiamiento propio, ahora incorporan un nuevo buque para sumar volumen a la línea de producción, pero fundamentalmente, calidad, a partir de la posibilidad de obtener una mayor proporción de langostino entero en condiciones óptimas de frescura.\nEl particular contexto que atraviesa la industria naval post Decreto 145 ocupó buena parte del discurso de Domingo Contessi. «Estamos a las puertas de un potencial resurgimiento de la Industria Naval de la mano de la pesca. Todos los astilleros del país nos sentimos comprometidos con este objetivo y dispuestos a invertir para estar a la altura de las circunstancias”, dijo el industrial en representación del Directorio.\nAntes de ingresar al astillero se observan los movimientos de obra en el nuevo predio gestionado ante el Consorcio Portuario donde planifican la ampliación para poder construir barcos más grandes que el Luca Santino.\nYa en el interior de la nave central se descubren en distintos grados de avance las nuevas construcciones planificadas para ser botadas este mismo año. La última, la 135, es otro barco para Cabo Vírgenes, el Espartano, un fresquero de 21 metros, también con innovaciones en su línea y equipamiento.\nTodo este movimiento generó que el astillero amplíe su planilla de personal un 60% y haya incorporado nuevas maquinarias para hacer frente a la mayor demanda de trabajo.\nLa ampliación se realiza sobre un terreno fiscal y con fondos propios. “Es difícil no pensar en estos momentos que hace apenas 6 años fuimos intimados a la restitución del predio por un interventor del Puerto de Mar del Plata”, dijo el directivo, en tanto reconoció que “ni aún en esos aciagos momentos se nos ocurrió protestar cortando los accesos al puerto o afectando el trabajo de otros, buscamos formas diferentes de visibilizar nuestro reclamo, litigamos judicialmente y gestionamos”, dijo Contessi.\nPareció un mensaje a los directivos de TC2 y la cooperativa Hipocoop, cuyos trabajadores y otros estibadores y transportistas, bloquearon todos los accesos la semana pasada y fue necesaria la intervención de Tizado.\n“Hoy debemos agradecer al Presidente y Directorio del Consorcio Portuario, porque después de 22 años de precariedad tenemos nuestro permiso de uso de suelo renovado y accedimos a un aumento de la superficie ocupada que nos permite concretar esta primera etapa de la ampliación edilicia”.\nContessi también planteó los peligros que merodean al desarrollo pleno de la industria naval. ”Hay sectores que pujan por una rebaja arancelaria para la importación de determinados segmentos de buques nuevos y usados. En la letra chica de esa medida, en una segmentación que no excluya a barcos como el de 40 metros que está en construcción dentro de nuestros galpones y en las excepciones absolutamente innecesarias a la normativa vigente, podría irse por la borda todo lo bueno que se ha hecho en estos últimos meses”, reconoció el Director.\nEntre las autoridades presentes en la botadura estuvieron el ministro de la Producción, Javier Tizado, quien destacó el trabajo del astillero y su apuesta por la generación de más empleo, inversión y desarrollo en tanto recordó el último bloqueo que sufrió el puerto y adujo problemas “de comunicación”. “El puerto es de ustedes”, les dijo a los presentes, entre los que había muchos actores pesqueros. “El sector de la pesca crece fuertemente y vamos a acompañar su desarrollo”, dijo el Ministro.\n    
#> 6 Tras un año de intento trabajo y una inversión de 3 millones de dólares, este jueves Astillero Contessi entregó su obra número 130. El buque fresquero de altura Luca Santino, ordenado por la pesquera Cabo Vírgenes, bien podría representar un mensaje para todos los actores involucrados en la industria pesquera.\nEl buque de 26,90 metros de eslora y doble cubierta cuenta con una máquina de hielo líquido que por primera vez se instala en un barco pesquero de la flota nacional. La innovación de Eduardo del Río fue destacada por Domingo Contessi desde el discurso previo a la botadura. Resignar bodega para instalar el dispositivo es un paso adelante hacia priorizar la calidad por sobre la cantidad.\nLa empresa armadora, encabezada por Del Río, reprocesa langostino en una planta de Rawson y hace poco adquirió dos buques fresqueros usados, casualmente, construidos en Contessi. Con financiamiento propio, ahora incorporan un nuevo buque para sumar volumen a la línea de producción, pero fundamentalmente, calidad, a partir de la posibilidad de obtener una mayor proporción de langostino entero en condiciones óptimas de frescura.\nEl particular contexto que atraviesa la industria naval post Decreto 145 ocupó buena parte del discurso de Domingo Contessi. «Estamos a las puertas de un potencial resurgimiento de la Industria Naval de la mano de la pesca. Todos los astilleros del país nos sentimos comprometidos con este objetivo y dispuestos a invertir para estar a la altura de las circunstancias”, dijo el industrial en representación del Directorio.\nAntes de ingresar al astillero se observan los movimientos de obra en el nuevo predio gestionado ante el Consorcio Portuario donde planifican la ampliación para poder construir barcos más grandes que el Luca Santino.\nYa en el interior de la nave central se descubren en distintos grados de avance las nuevas construcciones planificadas para ser botadas este mismo año. La última, la 135, es otro barco para Cabo Vírgenes, el Espartano, un fresquero de 21 metros, también con innovaciones en su línea y equipamiento.\nTodo este movimiento generó que el astillero amplíe su planilla de personal un 60% y haya incorporado nuevas maquinarias para hacer frente a la mayor demanda de trabajo.\nLa ampliación se realiza sobre un terreno fiscal y con fondos propios. “Es difícil no pensar en estos momentos que hace apenas 6 años fuimos intimados a la restitución del predio por un interventor del Puerto de Mar del Plata”, dijo el directivo, en tanto reconoció que “ni aún en esos aciagos momentos se nos ocurrió protestar cortando los accesos al puerto o afectando el trabajo de otros, buscamos formas diferentes de visibilizar nuestro reclamo, litigamos judicialmente y gestionamos”, dijo Contessi.\nPareció un mensaje a los directivos de TC2 y la cooperativa Hipocoop, cuyos trabajadores y otros estibadores y transportistas, bloquearon todos los accesos la semana pasada y fue necesaria la intervención de Tizado.\n“Hoy debemos agradecer al Presidente y Directorio del Consorcio Portuario, porque después de 22 años de precariedad tenemos nuestro permiso de uso de suelo renovado y accedimos a un aumento de la superficie ocupada que nos permite concretar esta primera etapa de la ampliación edilicia”.\nContessi también planteó los peligros que merodean al desarrollo pleno de la industria naval. ”Hay sectores que pujan por una rebaja arancelaria para la importación de determinados segmentos de buques nuevos y usados. En la letra chica de esa medida, en una segmentación que no excluya a barcos como el de 40 metros que está en construcción dentro de nuestros galpones y en las excepciones absolutamente innecesarias a la normativa vigente, podría irse por la borda todo lo bueno que se ha hecho en estos últimos meses”, reconoció el Director.\nEntre las autoridades presentes en la botadura estuvieron el ministro de la Producción, Javier Tizado, quien destacó el trabajo del astillero y su apuesta por la generación de más empleo, inversión y desarrollo en tanto recordó el último bloqueo que sufrió el puerto y adujo problemas “de comunicación”. “El puerto es de ustedes”, les dijo a los presentes, entre los que había muchos actores pesqueros. “El sector de la pesca crece fuertemente y vamos a acompañar su desarrollo”, dijo el Ministro.\n    
#>   id_token id_token_doc   token
#> 1        1            1    tras
#> 2        2            2      un
#> 3        3            3     año
#> 4        4            4      de
#> 5        5            5 intento
#> 6        6            6 trabajo