Colección de notas.

Contiene colecciones de notas de distintos portales noticiosos (una muestra corta). Una segunda colección es de notas del periódico bahiense La Nueva. También tiene resúmenes estadísticos de las bases completas para el desarrollo de los ejemplos de las funciones.

Usage

data(acep_bases)

Format

Es una lista con 8 objetos.

la_nueva

es una url para la descarga de una muestra del corpus de notas de La Nueva Provincia de Bahía Blanca

rev_puerto

es una url para la descarga del corpus de notas de la Revista Puerto

rp_procesada

es un data frame con indicadores de conflictividad basados en los datos de la Revista Puerto

lc_mdp

es una url para la descarga del corpus de notas de La Capital

rp_mdp

es una url para la descarga del corpus de notas de la Revista Puerto

ed_neco

es una url para la descarga del corpus de notas de Ecos Diarios

ln_bb

es una url para la descarga del corpus de notas de La Nueva

ln_arg

es una url para la descarga del corpus de notas de La Nación

spacy_postag

es un data.frame con una oración procesada con spacyr

titulares

es un vector con titulares de notas referidas a conflictos sociales

Source

Revista Puerto

La Nueva

References

Nieto, Agustín 2020 «Intersecciones entre historia digital e historia social: un ejercicio de lectura distante sobre la conflictividad marítima en la historia argentina reciente». Drassana: revista del Museu Maritim (28):122-42. (Revista Drassana)

Examples

acep_bases$rp_procesada[1:6, ]
#> # A tibble: 6 × 4
#>   fecha      n_palabras conflictos intensidad
#>   <date>          <int>      <int>      <dbl>
#> 1 2020-12-29         31          0     0     
#> 2 2020-12-28       1128          4     0.0035
#> 3 2020-12-24        530          0     0     
#> 4 2020-12-24        483          3     0.0062
#> 5 2020-12-23        525          1     0.0019
#> 6 2020-12-23        462          0     0