Limpieza de texto.

Función que limpia y normaliza las notas/textos.

Usage

acep_clean(
  x,
  tolower = TRUE,
  rm_cesp = TRUE,
  rm_emoji = TRUE,
  rm_hashtag = TRUE,
  rm_users = TRUE,
  rm_punt = TRUE,
  rm_num = TRUE,
  rm_url = TRUE,
  rm_meses = TRUE,
  rm_dias = TRUE,
  rm_stopwords = TRUE,
  rm_shortwords = TRUE,
  rm_newline = TRUE,
  rm_whitespace = TRUE,
  other_sw = NULL,
  u = 1
)

Arguments

x

vector de textos al que se le aplica la función de limpieza de texto.

tolower

convierte los textos a minúsculas.

rm_cesp

remueve caracteres especiales.

rm_emoji

remueve los emojis.

rm_hashtag

remueve los hashtags.

rm_users

remueve las menciones de usuarixs de redes sociales.

rm_punt

remueve la puntuación.

rm_num

remueve números.

rm_url

remueve las url.

rm_meses

remueve los meses del año.

rm_dias

remueve los días de la semana.

rm_stopwords

remueve palabras vacías.

rm_shortwords

remueve las palabras cortas.

rm_newline

remueve los saltos de línea.

rm_whitespace

remueve los espacios en blanco.

other_sw

su valor por defecto es NULL, sirve para ampliar el listado de stopwords con un nuevo vector de palabras.

u

umbral de caracteres para la función rm_shortwords.

Examples

rev_puerto <- acep_bases$rev_puerto
rev_puerto$nota_limpia <- acep_clean(rev_puerto$nota)
rev_puerto |> head()
#> # A tibble: 6 × 4
#>   fecha      titulo                                                nota  nota_…¹
#>   <date>     <chr>                                                 <chr> <chr>  
#> 1 2019-03-29 Astillero Contessi botó al Luca Santino, un fresquer… "Tra… tras a…
#> 2 2019-03-29 Di Leva despidió a trabajadores de El Marisco y Seba… "Tra… trabaj…
#> 3 2019-03-29 El Consejo aprobó la apertura al norte para el calam… "En … reunio…
#> 4 2019-03-28 “Todos los empresarios me dicen que tendrían que ach… "Cri… cristi…
#> 5 2019-03-28 En el Puerto de Montevideo siguen bajando un muerto … "La … organi…
#> 6 2019-03-28 Habilitan el muelle Piedra Buena para descarga de co… "La … admini…
#> # … with abbreviated variable name ¹​nota_limpia