Colección de diccionarios.

Colección de diccionarios que reúne diccionarios de diferentes orígenes. El diccionario dicc_confl_acep fueron construidos en el marco del Observatorio de Conflictividad de la UNMdP. Los diccionarios dicc_confl_gp y dicc_viol_gp fueron extraidos de Albrieu y Palazzo (2020).

data(acep_diccionarios)

Format

Es un objeto de clase 'list' con 3 componentes.

dicc_confl_gp

es un vector con palabras de un diciconario de términos que refeiren a conflictos

dicc_viol_gp

es un vector con palabras de un diciconario de términos que refeiren a violencia

dicc_confl_sismos

es un vector con palabras de un diciconario de términos que refeiren a conflictos

Source

Revista Puerto

La Nueva

References

Albrieu, Ramiro y Gabriel Palazzo 2020 «Categorización de conflictos sociales en el ámbito de los recursos naturales: un estudio de las actividades extractivas mediante la minería de textos». Revista CEPAL (131):29-59. (Revista de la CEPAL)

Laitano, Guillermina y Agustín Nieto «Análisis computacional de la conflictividad laboral en Mar del Plata durante el gobierno de Cambiemos». Ponencia presentado en VI Workshop - Los conflictos laborales en la Argentina del siglo XX y XXI: un abordaje interdisciplinario de conceptos, problemas y escalas de análisis, Tandil, 2021.

Examples

dicc_confl_acep <- acep_diccionarios$dicc_confl_sismos
dicc_confl_gp <- acep_diccionarios$dicc_confl_gp
dicc_confl_acep |> head()
#> [1] " twittazo"   " twitazo"    " semaforazo" " ladrillazo" " tractorazo"
#> [6] " banderazo" 
dicc_confl_gp |> head()
#> [1] "abatir"     "abdicación" "abdicar"    "abofetear"  "abolición" 
#> [6] "abolir"