Funcion para etiquetado POS, lematizacion, tokenizacion, extraccion de entidades.

Funcion que devuelve seis objetos data.frame con etiquetado POS (modelo spacyr) para su posterior procesamiento con la funcion acep_postag.

Usage

acep_postag(
  texto,
  core = "es_core_news_md",
  bajar_core = TRUE,
  inst_spacy = FALSE,
  inst_reticulate = FALSE
)

Source

Dependencias Universales para taggeo POS

Sobre el modelo spaCy de Python para R

Sobre el paquete tidygeocoder

Arguments

texto

vector con los textos a procesar.

core

idioma del modelo de etiquetado POS del paquete spacyr.

bajar_core

parametro booleano que define si bajar o no el modelo de etiquetado POS.

inst_spacy

parametro booleano que define si instalar o no spacy (python).

inst_reticulate

parametro booleano que define si instalar o no el paquete reticulate.

Value

Si todas las entradas son correctas, la salida sera una lista con tres bases de datos en formato tabular.

Examples

if (FALSE) {
texto <- "En Mar del Plata el SOIP declara la huelga en demanda de aumento salarial."
texto_postag <- acep_postag(texto)
texto_postag$texto_tag[ , c(1:8)]
}