Funcion para extraer triplets SVO (Sujeto-Verbo-Objeto).

Funcion que devuelve seis objetos data.frame con etiquetado POS (modelo spacyr) y relaciones sintacticas (modelo rsyntax) que permiten reconstruir estructuras sintacticas como SVO y Sujeto-Predicado. Una vez seleccionadas las notas periodisticas referidas a conflictos, esta funcion permite extraer sujetos de la protesta, accion realizada y objeto(s) de la accion. Tambien devuelve entidades nombradas (NER).

Usage

acep_svo(
  acep_tokenindex,
  prof_s = 2,
  prof_o = 2,
  relaciones = c("flat", "fixed", "appos"),
  conexiones = FALSE,
  rel_s = c("nsubj", "conj", "nmod"),
  rel_o = c("obj", "obl", "amod"),
  rel_evs = "nsubj",
  rel_evp = "obj",
  u = 1
)

Source

Dependencias Universales para taggeo POS

Sobre el paquete rsyntax

Arguments

acep_tokenindex

data.frame con el etiquetado POS y las relaciones de dependencia generado con la funcion acep_postag.

prof_s

es un numero entero positivo que determina la profundidad a la que se buscan las relaciones dentro del sujeto. Este parametro se hereda del la funcion children() del paquete rsyntax. Se recomienda no superar el valor 2.

prof_o

es un numero entero positivo que determina la profundidad a la que se buscan las relaciones dentro del objeto. Este parametro se hereda del la funcion children() del paquete rsyntax. Se recomienda no superar el valor 2.

relaciones

vector con las etiquetas de relaciones que se usaran en la funcion de relleno. Este parametro se hereda de la funcion custom_fill() del paquete rsyntax.

conexiones

es un parametro heredado del paquete rsyntax que controla el comportamiento si la profundidad es > 1 y se utilizan filtros. Si es FALSE, se recuperan todos los padres/hijos hasta la profundidad dada, y luego se filtran. De esta manera, los nietos que satisfacen las condiciones del filtro se recuperan incluso si sus padres no satisfacen las condiciones. Si es TRUE, el filtro se aplica en cada nivel de profundidad, de modo que solo se recuperan las ramas totalmente conectadas de los nodos que satisfacen las condiciones. Este parametro se hereda de la funcion custom_fill() del paquete rsyntax.

rel_s

vector de etiquetas de relaciones sintacticas en el sujeto. Este parametro se hereda del la funcion children() del paquete rsyntax.

rel_o

vector de etiquetas de relaciones sintacticas en el objeto. Este parametro se hereda del la funcion children() del paquete rsyntax.

rel_evs

etiqueta de relaciones a ser agregada en la reconstruccion del sujeto.

rel_evp

etiqueta de relaciones a ser agregada en la reconstruccion del predicado.

u

numero entero que indica el umbral de palabras del objeto en la reconstruccion SVO.

Value

Si todas las entradas son correctas, la salida sera una lista con tres bases de datos en formato tabular.

References

Welbers, K., Atteveldt, W. van, & Kleinnijenhuis, J. 2021. Extracting semantic relations using syntax: An R package for querying and reshaping dependency trees. Computational Communication Research, 3-2, 1-16. (link al articulo)

Examples

texto <- acep_postag("El SUTEBA fue al paro por mejoras salariales.")
#> Installing model "es_core_news_md"
#> Error: Error 1 occurred installing packages into conda environment spacy_condaenv
acep_svo(texto$texto_tag)
#> Error in methods::is(tokens, "tokenIndex"): object 'texto' not found