Funcion para extraer triplets SVO (Sujeto-Verbo-Objeto).

Funcion que devuelve cinco objetos data.frame con etiquetado POS (modelo udpipe) y relaciones sintacticas (modelo rsyntax) que permiten reconstruir estructuras sintacticas como SVO y Sujeto-Predicado. Una vez seleccionadas las notas periodisticas referidas a conflictos, esta función permite extraer sujetos de la protesta, acción realizada y objeto(s) de la acción.

Usage

acep_svo(
  texto,
  modelo = "spanish",
  prof_s = 2,
  prof_o = 2,
  relaciones = c("flat", "fixed", "compount"),
  conexiones = FALSE,
  rel_s = c("nsubj", "conj"),
  rel_o = c("obj", "obl", "nmod"),
  rel_evs = "nsubj",
  rel_evp = "obj",
  u = 1
)

Arguments

texto

vector con los textos a procesar.

modelo

idioma del modelo de etiquetado POS del paquete udpipe.

prof_s

profundidad de relaciones sintacticas para el sujeto.

prof_o

profundidad de relaciones sintacticas para el objeto.

relaciones

vector con las relaciones de analisis sintactico del paquete rsyntax.

conexiones

parametro logico (TRUE/FALSE) para la reconstruccion de las conexiones sintacticas.

rel_s

vector de etiquetas de relaciones sintacticas en el sujeto.

rel_o

vector de etiquetas de relaciones sintacticas en el objeto.

rel_evs

etiqueta de relaciones a ser agregada en la reconstruccion del sujeto.

rel_evp

etiqueta de relaciones a ser agregada en la reconstruccion del objeto.

u

numero entero que indica el umbral de palabras del objeto en la reconstruccion SVO.

Value

Si todas las entradas son correctas, la salida sera una lista con tres bases de datos en formato tabular.

Examples

texto <- "El SOIP declara la huelga en demanda de aumento salarial."
texto_svo <- acep_svo(texto)
#> Downloading udpipe model from https://raw.githubusercontent.com/jwijffels/udpipe.models.ud.2.5/master/inst/udpipe-ud-2.5-191206/spanish-gsd-ud-2.5-191206.udpipe to /home/runner/work/ACEP/ACEP/docs/reference/spanish-gsd-ud-2.5-191206.udpipe
#>  - This model has been trained on version 2.5 of data from https://universaldependencies.org
#>  - The model is distributed under the CC-BY-SA-NC license: https://creativecommons.org/licenses/by-nc-sa/4.0
#>  - Visit https://github.com/jwijffels/udpipe.models.ud.2.5 for model license details.
#>  - For a list of all models and their licenses (most models you can download with this package have either a CC-BY-SA or a CC-BY-SA-NC license) read the documentation at ?udpipe_download_model. For building your own models: visit the documentation by typing vignette('udpipe-train', package = 'udpipe')
#> Downloading finished, model stored at '/home/runner/work/ACEP/ACEP/docs/reference/spanish-gsd-ud-2.5-191206.udpipe'
texto_svo$acep_list_svo
#> # A tibble: 1 × 4
#>   doc_id parrafo_id oracion_id eventos                  
#>    <int>      <int>      <int> <chr>                    
#> 1      1          1          1 SOIP -> declara -> huelga